# .specify/specs/024-donor-bidding-ui/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Donor Bidding UI API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /watchlist:
    get:
      summary: Get donor watch list
      responses:
        "200":
          description: Watch list items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchListResponse"
    post:
      summary: Add item to watch list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchListRequest"
      responses:
        "201":
          description: Watch list entry created
  /watchlist/{item_id}:
    delete:
      summary: Remove item from watch list
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Watch list entry removed
...
components:
  schemas:
...
    AuctionItemSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        current_bid_amount:
          type: number
        min_next_bid_amount:
          type: number
        bid_count:
          type: integer
        bidding_open:
          type: boolean
        watcher_count:
          type: integer
        buy_now_enabled:
          type: boolean
        buy_now_remaining_quantity:
          type: integer
        promotion_badge:
          type: string
        promotion_notice:
          type: string
...
    WatchListRequest:
      type: object
      required: [item_id]
      properties:
        item_id:
          type: string
    WatchListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuctionItemSummary"
...